refactor streamable http service
