+++ src/mcp/index.ts
@@ -5,3 +5,5 @@ export * from './services/mcp-registry.service';
 export * from './services/mcp-executor.service';
 export * from './services/mcp-sse.service';
 export * from './services/mcp-streamable-http.service';
+export * from './services/mcp-streamable-http-stateless.service';
+export * from './services/mcp-streamable-http-stateful.service';
